+++ cuda/tests/test_cuda.cu
@@ -11,9 +11,9 @@
 // Device/constant memory definitions required by cuda_scene.h (and cuda_color.h)
 // ---------------------------------------------------------------------------
 __constant__ double cuda::d_flux_lut[cuda::MAX_FLUX_LUT_ENTRIES];
-__device__ cuda::Star cuda::d_stars[cuda::MAX_STARS];       // global (too large for __constant__)
-__constant__ double cuda::d_color_lut[cuda::MAX_SPECTRUM_ENTRIES][3];
-__constant__ double cuda::d_luminosity_lut[cuda::MAX_SPECTRUM_ENTRIES];
+__device__ cuda::Star cuda::d_stars[cuda::MAX_STARS];       // global (60KB with float, exceeds cmem budget)
+__constant__ float cuda::d_color_lut[cuda::MAX_SPECTRUM_ENTRIES][3];
+__constant__ float cuda::d_luminosity_lut[cuda::MAX_SPECTRUM_ENTRIES];
 
 // ---------------------------------------------------------------------------
 // Device kernel: runs 7 independent math sub-tests, writing pass/fail (1/0)
